Move templates package out of api directory
The idea behind this change is to remove non-api code from apis/contrail directory.
This directory is used by tools like operator-sdk or kubebuilder to generate code.
The fact that we have some additional logic makes it harder to migrate to newer versions of those tools.
