+++ pkg/apis/contrail/v1alpha1/BUILD.bazel
@@ -38,8 +38,8 @@ go_library(
     importpath = "github.com/Juniper/contrail-operator/pkg/apis/contrail/v1alpha1",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/apis/contrail/v1alpha1/templates:go_default_library",
         "//pkg/certificates:go_default_library",
+        "//pkg/configuration:go_default_library",
         "@com_github_go_openapi_spec//:go_default_library",
         "@in_gopkg_yaml.v2//:go_default_library",
         "@io_k8s_api//apps/v1:go_default_library",
