+++ pkg/configuration/BUILD.bazel
@@ -13,7 +13,7 @@ go_library(
         "webui_config.go",
         "zookeeper_config.go",
     ],
-    importpath = "github.com/Juniper/contrail-operator/pkg/apis/contrail/v1alpha1/templates",
+    importpath = "github.com/Juniper/contrail-operator/pkg/configuration",
     visibility = ["//visibility:public"],
     deps = ["@io_k8s_api//core/v1:go_default_library"],
 )
