+++ pkg/configuration/control_config.go
@@ -1,4 +1,4 @@
-package templates
+package configuration
 
 import "text/template"
