+++ pkg/configuration/templates_helper_functions.go
@@ -1,4 +1,4 @@
-package templates
+package configuration
 
 import (
 	"strconv"
