+++ pkg/configuration/templates_helper_functions_test.go
@@ -1,4 +1,4 @@
-package templates
+package configuration
 
 import (
 	"testing"
